Compute postdominance correctly for infinite loop
Bias *predecessor* traversal root finding so that you use
a later block in the original list.  This ensures that
for certain simple infinite loops in the CFG where neither
block branches to a node without successors, that we'll
compute the loop header as dominating the latch block, and the
latch block as postdominating the loop header.