Don't use friendly names for SPIR-V validation errors
Setting up friendly names requires parsing the entire module
which costs a lot of time due to (1) scanning the whole module,
and (2) creating string names for many IDs.

The validation errors are only emitted when there is a compiler
bug, so they should be rare.  Don't make every compilation pay
the cost.